+++ src/main/java/uk/org/binky/lexinator/Tokenizer.java
@@ -1,0 +1,5 @@
+package uk.org.binky.lexinator;
+
+public interface Tokenizer<T extends Enum<T>> {
+    Token<T> getToken();
+}