+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 - [x] Lexer.setState is no longer protected, but private
 - [x] Lexer's constructors are not protected
 - [x] Removed deprecated whitespace methods
+- [x] Add Tokenizer interface
+- [x] Pulled expect out into a seperate class, now usable for Tokenizer
 
 ## v0.1.4